==> telegram/00_tg_ticker_in_msg_tg.ipynb <==
## Данный ноутбук ищет упоминания о компаниях в сообщениях

import pandas as pd
from tqdm import tqdm_notebook as tqdm

### Готовим словарь тикеров

tickers = {
'gazp' : ['gazp', 'газпром',   'миллер', 'газик', 'нашевсе', 'северный поток', 'южный поток', 'сила сибири', \
          'ru0007661625'],
    
'sber' : ['sber', 'сбер', 'греф', 'ru0009029540'], # добавить названия активов
'vtbr' : ['vtb', 'костин', 'втб', 'внешторгбанк', 'народное ipo', 'ru000A0JP5V6'],
'rosn':  ['rosn', 'роснефт', 'сечин', 'ru000a0j2q06'],
'gmkn':  ['gmkn', 'норникел', 'nornickel', 'потанин', 'гмк', 'норк', 'интеррос', 'гамак', 'ru0007288411'],
'moex':  ['московская биржа', 'мосбирж', 'moex', 'москухн', 'ru000a0jr4a1'],
'alrs':  ['алрос', 'alrosa', 'alrs', 'алмаз', 'бриллиант', 'нюрб', 'ru0007252813'],
'lkoh':  ['lkoh', 'luk', 'алекперов', 'лукойл', 'лучок', 'лучк', 'ru0009024277'],
'sngs':  ['сургут', 'sngs', 'surgutneftega', 'кубышка', 'сурик', 'снгс', 'ru0008926258'],
'tatn':  ['татнефт', 'tatneft', 'tatn', 'маганов', 'ru0009033591'],
'yndx':  ['yndx', 'yandex', 'яндекс', 'волож'],
'aflt':  ['аэрофлот', 'aflt', 'добролет', 'аврор', 'российские авиалинии', 'полубояринов'],
'fees':  ['фск еэс', 'fees', 'феес', 'федеральные электросети', 'феск'],
'magn':  ['ммк', 'magn', 'магнитк', 'магнитогорск', 'магна'],
'mtss':  ['mts', 'мтс'],
'nvtk':  ['nvtk', 'новатэк', 'нова тэк', 'novatek', 'ямал спг', 'михельсон'],
'nlmk':  ['нлмк', 'nlmk', 'липецк', 'лисин', 'липк'],
'chmf':  ['северстал', 'chmf', 'severstal', 'севка', 'череповец', 'мордашов'],
'mgnt':  ['mgnt', 'магнит', 'magnit', 'марафон групп', 'галицкий', 'тандер'],
'plzl':  ['plzl', 'polyus', 'полюс', 'polyusgold', 'сухой лог'],
'qiwi':  ['qiwi', 'киви'],
'rasp':  ['rasp', 'распад'],
'lsrg':  ['лср', 'lsr', 'элэсэр'],
'mtlr':  ['мечел', 'mtlr', 'mechel', 'ru000a0jpv70'],
'phor':  ['фосагр', 'phor', 'phosagr', 'фосфат', 'dap'],
'rtkm':  ['rtkm', 'rostelecom', 'ростел', 'tele2', 'ртк', 'рткм', 'осеевский'],
'hydr':  ['русгидр', 'hydr', 'rushydr', 'гидра', 'ru000a0jpkh7'],
'afks':  ['афк', 'afks', 'afk sistem', 'ozon', 'евтушенков', 'систем'],
'trmk':  ['тмк', 'трубная', 'trmk', 'tmk', 'трмк'],
'upro':  ['юнипро', 'upro', 'unip', 'эон россия', 'березовская грэс'],
'akrn':  ['акрон', 'akrn', 'akron', 'азотные удобрения'],
'pikk':  ['пик', 'pik', 'мортон'],
'lnta':  ['лент', 'lent', 'lnta', 'tpg capital'],
'msng':  ['мосэнерг', 'msng', 'mosenerg', 'мося'],
'tgka':  ['тгк1', 'тгк-1', 'тгк 1', 'tgka', 'tgk-1', 'tgk1', 'tgk 1'],
'rual':  ['русал', 'rusal', 'rual', 'русский алюминий', 'краз'],
'trnf':  ['транснефт', 'trnf', 'нефтепровод дружба'],
'sibn':  ['газпромнефт', 'sibn', 'gazpromneft', 'sibneft', 'сибнефт', 'газнефт'],
'nmtp':  ['нмтп', 'nmtp', 'новороссийск', 'новопорт'],
'tanl':  ['tanl', 'tantal', 'тант'],
'iskj':  ['искч', 'институт стволовых клеток', 'стволовые клетки', 'стволовых клеток', 'iskj'],
'yakg':  ['ятэк', 'yakg', 'yakutsk fue', 'якутская топл', 'ru0007796819'],
'gtlc':  ['gtl', 'джитиэл', 'гтл'],
'gche':  ['черкизов', 'cherkizov', 'che'],
'rolo':  ['русолов', 'rolo', 'rusolov', 'ролово'],
'svav':  ['соллерс', 'svav', 'sollers', 'северстальавт', 'уаз'],
'arsa':  ['арсагер', 'ук арса', 'arsa', 'arsager'],
'krkn':  ['саратовский нпз', 'снпз', 'krkn', 'сарнпз'],
'selg':  ['селигдар', 'seligdar', 'selg'],
'prfn':  ['чзпсн', 'профнастил', 'настил', 'prfn'],
'gaza':  ['gaza', 'газ завод', 'газ авто', 'газель', 'горьковский'],
'kbtk':  ['kbtk', 'кбтк', 'кузбас', 'даэния холдинг'],
'nknc':  ['nknc', 'нижнекамск', 'нкнх', 'нефтехим'],
'aptk':  ['aptk', 'аптек', '36 и 6', '36.6'],
'amez':  ['amez', 'ашинский', 'аша', 'amet', 'амез', 'амет'],
'dasb':  ['dasb', 'даг'],
'irkt':  ['иркут', 'irkt', 'сухой', 'суперджет', 'sukhoi', 'ssj'],
'lsng':  ['ленэнерг', 'lsng', 'ленк', 'петербургские электросети'],
'rlmn':  ['ролл', 'ролман', 'rlmn', 'роля'],
'obuv':  ['обувь росси', 'obuv', 'westfalika', 'обувьрус'],
'abrd':  ['абрау', 'дюрсо', 'abrd', 'abrau', 'durso', 'русский винный дом', 'борис титов', 'ru000a0js5t7'],
'agro':  ['agro', 'русагро', 'us7496552057'],
'albk':  ['бест эффортс', 'albk', 'best efforts', 'алор банкбестэфб', 'ru000a0js2m9'],
'alnu':  ['alnu', 'nyurba', 'алроса-нюрба', 'якуталмаз', 'ru000a0jp468'],
'aqua':  ['аквакультураaqua', 'aquacultur', 'русаква', 'русское море', 'ru000a0jqts3'],
'assb':  ['астрэнсб', 'assb', 'astrakhan', 'аэск', 'астраханэнергосбыт', 'ru000a0d8mm8'],
'avan':  ['avangard', 'авангард', 'ru000a0dm7b3'],
'bane':  ['башнефт', 'bane', 'bashneft', 'ru0007976957'],
'belu':  ['белуг', 'belu', 'belug', 'ru000a0hl5m1'],
'bisv':  ['bisv', 'bashinformsvyaz', 'bashtel', 'ru0009059216'],
'blng':  ['белон', 'blng', 'belon', 'ru000a0j2qg8'],
'brzl':  ['бурзолот', 'brzl', 'buryatzolot', 'бурятзолот', 'ru0009288658'],
'bspb':  ['бсп', 'bspb', 'bank st petersburg', 'банк санкт-петербург', 'ленбанк', 'ru0009100945'],
'cbom':  ['мкб', 'cbom', 'credit bank of moscow', 'московский кредитный банк', 'ru000a0jug31'],
'chep':  ['чтпз', 'chep', 'chelyabinsk pipe plant', 'chelpipe', 'chtpz', 'челябинский трубопрокатн', 'ru0009066807'],
'chgz':  ['рн-запсиб', 'рн запсиб', 'chgz', 'rn-western siberia', 'рн-западная сибирь', 'буй', 'ru0009082291'],
'chkz':  ['чкпз', 'chkz', 'chkpz', 'челябинский кузнечно-прессовый заво', 'chelyabinsk forge & press plant', \
          'гартунг', 'ru000a0jnst0'],
'cntl':  ['телеграф', 'cntl', 'telegraf', 'ru0007665147'],
'diod':  ['заводдиод', 'diod', 'диод', 'ru000a0jqwc1'],
'dsky':  ['dsky', 'detskiy mir', 'детский', 'ru000a0jsq90'],
'dvec':  ['дэк', 'dvec', 'feec', 'far-eastern energy', 'дальневосточная энергетическая компания', 'ru000a0jp2w1'],
'dzrd':  ['донскзр', 'dzrd', 'donskoy factory of radiocomponents', 'донской завод радиодеталей', \
          'donskoy zavod radiodetaley', 'ru000a0jqw11'],
'eelt':  ['евроэлтех', 'eelt', 'european eltech', 'европэлектротехника', 'европейская электротехника', 'каленков', \
          'ru000a0jww54'],
'eltz':  ['eltz', 'electrozinc', 'электроцинк', 'ru000a0b66v1'],
'enpg':  ['эн+', 'enpg', 'en plus', 'en+', 'ru000a100k72'],
'enru':  ['энелрос', 'enru', 'enel', 'энел', 'ru000a0f5un3'],
'etln':  ['etln', 'etalon', 'эталон', 'us29760g1031'],
'fesh':  ['двмп', 'fesh', 'fesco', 'дв морское пароходство', 'far-eastern shipping co plc', 'ru0008992318'],
'five':  ['five', 'x5', 'пятерочк', 'шехтерман', 'перекрест', 'карусел', 'us98387e2054'],
'flot':  ['совкомфлот', 'flot', 'sovcomflot', 'современный коммерческий флот', 'ru000a0jxnu8'],
'gazc':  ['газкон', 'gazc', 'ru000a0jqg86'],
'gazs':  ['газсервис', 'газ-сервис', 'gazs', 'gaz-service', 'ru000a0jqg78'],
'gazt':  ['газтек', 'газ-тек', 'gazt', 'gaz-tek', 'ru000a0js199'],
'gema':  ['ммцб', 'gema', 'mmtsb', 'biomaterials processing and cryostorage', 'гема', \
          'обработки и криохранения биоматериалов', 'ru000a100gc7'],
'gltr':  ['gltr', 'глобалтранс', 'globaltrans', 'us37949e2046'],
'gtrk':  ['гтм', 'gtrk', 'globaltruck', 'глобалтрак', 'gtm', 'ru000a0zyd22'],
'gtss':  ['геотек', 'gtss', 'geotech', 'сейсморазведка', 'ru000a0jq7u3'],
'hals':  ['галс', 'hals', 'пучков', 'ru000a0jnp96'],
'himc':  ['химпром', 'himc', 'khimprom', 'ru0009098990'],
'igst':  ['ижстал', 'igst', 'izhstal', 'козеннов', 'ru0002155292'],
'irao':  ['irao', 'inter rao', 'борис ковальчук', 'ru000a0jpnm1', 'рао', 'интер'],
'irgz':  ['иркэнерго', 'irgz', 'irkutskenergo', 'иркутскэнерго', 'ru0008960828'],
'jnos':  ['славн-янос', 'jnos', 'slavneft yaroslavnefteorgsintez', 'славнефт', 'ярославнефт', 'ru0009100895'],
'kazt':  ['куйбазот', 'kazt', 'kuibyshevazot', 'куйбышевазот', 'ru000a0b9bv2'],
'kbsb':  ['тнсэкубань', 'kbsb', 'tns energo kuban', 'энерго кубань', 'кубанская энергосбытовая компания', 'ru000a0jnj11'],
'kche':  ['камчатскэ', 'kche', 'kamchatskenergo', 'ru0006753498'],
'kgkc':  ['кургангк', 'kgkc', 'kurgan generation', 'курганская генер', 'кгк', 'ru000a0jp120'],
'klsb':  ['калужскск', 'klsb', 'kaluga power', 'калужская сбыт', 'ru000a0dkzk3'],
'kmaz':  ['камаз', 'kmaz', 'kamaz', 'ru0008959580'],
'kmez':  ['кмз', 'kmez', 'kovrov', 'ковровск', 'ru0006753613'],
'kmtz':  ['косогмз', 'kmtz', 'kosaya gora', 'косогорск', 'kosogorski', 'ru000a0hf1l0'],
'kogk':  ['коршгок', 'kogk', 'korshynov', 'коршуновск', 'ru0002155359'],
'krko':  ['ткзкк', 'krko', 'krasny kotelshchik', 'красный котельщик', 'ru0009098123'],
'krot':  ['красокт', 'krot', 'крот', 'krasnyj octyabr', 'красный октябрь', 'krasny oktyabr', 'ru0008913850'],
'krsb':  ['красэсб', 'krsb', 'krasnoyarskenergo', 'красноярскэнерго', 'ru000a0hmly1'],
'ksgr':  ['кокс', 'ksgr', 'koks', 'ru000a0ernt3'],
'ktsb':  ['ксб', 'ktsb', 'kostroma', 'костромск', ' kostromskaya sbyto', 'костромская сбытовая компания', 'ru000a0d8lw9'],
'kube':  ['кубанэн', 'kube', 'kubanenergo', 'кубаньэнерго', 'kuban power', 'rosseti kuban', 'ru0009046767'],
'kunf':  ['кузоцм', 'kunf', 'kuzocm', 'каменск-уральский', 'каменск уральский', 'ru000a0jpge2'],
'kuzb':  ['кузнецкий банк', 'kuzb', 'bank kuznetskiy', 'kuznetskiy bank', 'банк кузнец', 'ru000a0jsq66'],
'kzms':  ['кзмс', 'kzms', ' krasnokamsk metal', 'краснокамский завод', 'ru000a0jnhk2'],
'kzos':  ['органический синтез', 'kzos', 'оргсинтез', 'organicheskiy sintez', 'ru0009089825'],
'life':  ['фармсинтез', 'life', 'pharmsynt', 'ru000a0jr514'],
'lnzl':  ['лензолото', 'lnzl', 'lenzoloto', 'ленское золотодобывающ', 'ru000a0jp1n2'],
'lpsb':  ['лэск', 'lpsb', 'липецкая энерго', 'lipetsk power', 'lipetsk energ', 'ru000a0d8mr7'],
'mage':  ['магадэн', 'mage', 'magadanen', 'магаданэн', 'ru0006758919'],
'mail':  ['mail', 'mail.ru', 'mail', 'мэйл.ру', 'мэйл ру', 'добродеев', 'us5603172082'],
'mfgs':  ['мегион', 'mfgs', 'megionneftegas', 'ru0009011126'],
'mgnz':  ['смз', 'mgnz', 'smz', 'соликамский магн', 'ru0009100911'],
'mgts':  ['мгтс', 'mgts', 'moscow city tele', 'московская городская телефонная', 'moskovskaya gorodskaya tele',\
          'ru0009036461'],
'misb':  ['тнсэнмарэ', 'misb', 'tns energo mari', 'тнс энерго мари', 'ru000a0d8k33'],
'mobb':  ['mobb', 'moskovskii oblastnoi bank', 'московский областной банк', 'мособлбанк', 'ru000a0js7v9'],
'mori':  ['морион', 'mori', 'morion', 'телекоммуникационное оборудование', 'ru0009092431'],
'mrkc':  ['мрск центр', 'mrkc', 'ru000a0jppl8'],
'mrkk':  ['россети ск', 'mrkk', 'rosseti north caucasus', 'мрск северного кавказа', 'ru000a0jppq7'],
'mrkp':  ['мрск цп', 'mrkp', 'мрск центра и приволжья', 'ru000a0jpn96'],
'mrks':  ['mrks', 'rosseti siberia', 'россети сибирь', 'mrsk sibiri', 'мрск сибири', 'ru000a0jppf0'],
'mrku':  ['мрск ур', 'mrku', 'idgc of ural', 'ru000a0jppt1'],
'mrkv':  ['mrkv', 'rosseti volg', 'россети волг', 'мрск волг', 'ru000a0jppn4'],
'mrky':  ['россюг', 'mrky', 'rosseti south', 'россети юг', 'мрск юг', 'ru000a0jppg8'],
'mrkz':  ['мрск сз', 'mrkz', 'мрск северо-запад', 'mrsk severo zapad', 'ru000a0jppb9'],
'mrsb':  ['мордэнсб', 'mrsb', 'mordovia energ', 'мордовская энерг', 'ru000a0d9aj7'],
'msrs':  ['msrs', 'rosseti mosc', 'россети моск', 'московская объедин', 'ru000a0et7y7'],
'mstt':  ['мостотрест', 'mstt', 'mostotrest', 'ru0009177331'],
'mvid':  ['мвидео', 'mvid', 'м.видео', 'м видео', 'эльдорадо', 'изосимов', 'ru000a0jpga0'],
'nauk':  ['нпонаука', 'nauk', 'нпо наука', 'ru000a0jrpx9'],
'nfaz':  ['нефтекамск', 'nfaz', 'nefaz', 'ru0009115604'],
'nkhp':  ['нкхп', 'nkhp', 'novorossiysk grain', 'novorossiyskiy kombi', 'новороссийский комбин', 'ru000a0blwd7'],
'nksh':  ['нижкамшина', 'nksh', 'nizhnekamskshina', 'нижнекамскшина', 'ru0009100515'],
'nnsb':  ['тнсэннн', 'nnsb', 'tns energo n', 'тнс энерго н', 'ru000a0et5a1'],
'npof':  ['физик', 'npof', 'fizik', 'ru000a0jq5x1'],
'nsvz':  ['наукасвяз', 'nsvz', 'nauka-svyaz', 'nauka-tele', 'наука-связ', 'наука связ', 'ru000a0jqlb6'],
'ogkb':  ['огк-2', 'огк2', 'ogkb', 'ogk-2', 'семиколенов', 'ru000a0jng55'],
'paza':  ['павлавт', 'paza', 'pavlovo', 'павловский авт', 'pavlovskiy avt', 'паз', 'ru0009083240'],
'pmsb':  ['пермьэнсб', 'pmsb', 'perm energ', 'пермэнерг', 'permskaya energ', 'пермская энерг', 'ru000a0et123'],
'prmb':  ['prmb', 'primorie bank', 'приморье банк', 'банк приморье', 'ru000a0dpnq5'],
'prtk':  ['протек', 'prtk', 'protek', 'ru000a0jqu47'],
'rbcm':  ['рбк', 'rbcm', 'rbk', 'rbc', 'ru000a0jr6a6'],
'rdrb':  ['рдбанк', 'rdrb', 'rosdorbank', 'российскийдор', 'российский дор', 'ru000a0ju6q7'],
'rgss':  ['ргс', 'rgs', 'rosgosstra', 'росгосстра', 'гальперин', 'ru0008010855'],
'rkke':  ['энергияркк', 'rkke', 'spc energia', 'ркк энергия', 'ru0009095939'],
'rnft':  ['rnft', 'russneft', 'русснефт', 'ru000a0jse60'],
'rosb':  ['росбанк', 'rosb', 'rosbank', 'ru000a0hhk26'],
'rsti':  ['россет', 'rsti', 'российские сети', 'ru000a0jpvj0'],
'rtgz':  ['rtgz', 'rostovoblgaz', 'ростовоблгаз', 'ru000a0jpg04'],
'rtsb':  ['rtsb', 'tns energo rost', 'тнс энерго рост', 'ru000a0d8pb4'],
'rugr':  ['русгрэйн', 'ругрэйн', 'rugr', 'rusgrain', 'русское зерн', 'ru000a0jpnp4'],
'rusi':  ['икрусс', 'rusi', 'руссинвест', 'рус инвест', 'ru000a0jq9w5'],
'rusp':  ['русполимет', 'rusp', 'ruspol', 'ru000a0jnh21'],
'rzsb':  ['rzsb', 'ryazan energ', 'рязанская энерг', 'рязаньэнерго', 'ru000a0d9af5'],
'sago':  ['sago', 'samaraen', 'самараэнерго', 'самарэнерго', 'ru0009098255'],
'sare':  ['sare', 'saratoven', 'саратовэнерго', 'ru0009100754'],
'sfin':  ['сафмар', 'sfin', 'safmar', 'европлан', 'ru000a0jvw89'],
'slen':  ['сахэнер', 'slen', 'sakhalinen', 'сахалинэн', 'ru0009280465'],
'smlt':  ['самолет', 'smlt', 'samolet', 'ru000a0zzg02'],
'stsb':  ['stsb', 'stavropolen', 'ставрополэн', 'ставропольэн', 'ru000a0et1w4'],
'tasb':  ['tasb', 'tambov en', 'тамбовская эн', 'тамбовэнерго', 'ru000a0d8ng8'],
'tcsg':  ['tcs', 'tinkof', 'тисиэс', 'тинек', 'тинько', 'us87238u2033'],
'tgkb':  ['тгк 2', 'тгк-2', 'тгк2', 'tgkb', 'tgc-2', 'tgc2', 'tgc 2', 'ru000a0jngs7'],
'tgkd':  ['квадр', 'tgkd', 'quadra', 'ru000a0jnmz0'],
'tgkn':  ['тгк-14', 'тгк14', 'тгк 14', 'tgkn', 'tgk-14', 'tgk14', 'tgk 14', 'ru000a0h1es3'],
'tnse':  ['тнсэнерг', 'tns', 'tns energ', 'тнс энерг', 'ru000a0jucq5'],
'tors':  ['трк', 'tors', 'tomsk distribution', 'томская распред', 'ru000a0etzf2'],
'trcn':  ['транск', 'trcn', 'transcontainer', 'трансконтейнер', 'транс контейнер', 'ru000a0jprx9'],
'trfm':  ['трансф', 'trfm', 'transfin', 'трансфин', 'ru000a0jugp8'],
'ttlk':  ['таттел', 'ttlk', 'tattel', 'ru000a0hm5c1'],
'tuza':  ['тза', 'tuza', 'tuimazinsk', 'туймазинск', 'teuza', 'ru000a0hl7a2'],
'ucss':  ['окс', 'ucss', 'united credit', 'объединенные кредитн', 'ru000a0jtqk0'],
'ukuz':  ['южкузб', 'ukuz', 'southern kuzb', 'южный куз', 'юж куз', 'ru0005294775'],
'unac':  ['авиасткао', 'unac', 'united aircraft', 'объединенная авиа', 'ru000a0jplz7'],
'unkl':  ['юунк', 'unkl', 'southern ural', 'южно-уральский ник', 'ru0004887991'],
'urka':  ['урка', 'urka', 'uralka', 'уралка'],
'urkz':  ['уркузн', 'urkz', 'urals stampings', 'уралкуз', 'уральская кузниц', 'ru000a0jpfy2'],
'usbn':  ['уралсиб', 'usbn', 'uralsib', 'урсиб', 'ru0006929536'],
'utar':  ['ютэйр', 'utar', 'utair', 'ru0007661385'],
'uwgn':  ['овк', 'uwgn', 'united wagon', 'нпк овк', 'ru000a0jvbt9'],
'vgsb':  ['vgsb', 'volgograden', 'волгоградэн', 'volgograd en', 'ru000a0d8l73'],
'vjgz':  ['варьеган', 'vjgz', 'varyogan', 'varyegan', 'ru0009046460'],
'vlhz':  ['вхз', 'vlhz', 'vladimir chem', 'владимирский хим', 'ru0007984761'],
'vrsb':  ['vrsb', 'tns energo voronez', 'тнс энерго воронеж', 'ru000a0dpg67'],
'vsmo':  ['всмпо-авсм', 'vsmo', 'vsmpo-avisma', 'ависма', 'ru0009100291'],
'vsyd':  ['выбсуд', 'vsyd', 'vyborg ship', 'vyborgskiy sud', 'выборгский суд', 'ru0006334372'],
'vzrz':  ['возрожд', 'vzrz', 'vozrozhd', 'ru0009084214'],
'wtcm':  ['цмт', 'wtcm', 'world trade cen', 'центр международной торговли', 'ru0008137070'],
'yken':  ['якутскэнерг', 'yken', 'yakutskenerg', 'ru0009257075'],
'yrsb':  ['yrsb', 'tns energo yarosavl', 'тнс энерго ярославл', 'ru000a0d8760'],
'zill':  ['зил', 'zill', 'likhach', 'лихачев', 'ru0009086193'],
'zvez':  ['звезд', 'zvez', 'zvezd', 'ru0009091300']
}

#### Словарь тикеров у которых новостей больше 200 за весь период

tickers = {'vtbr': ['vtb',
  'костин',
  'втб',
  'внешторгбанк',
  'народное ipo',
  'ru000A0JP5V6'],
 'zvez': ['звезд', 'zvez', 'zvezd', 'ru0009091300'],
 'zill': ['зил', 'zill', 'likhach', 'лихачев', 'ru0009086193'],
 'yndx': ['yndx', 'yandex', 'яндекс', 'волож'],
 'uwgn': ['овк', 'uwgn', 'united wagon', 'нпк овк', 'ru000a0jvbt9'],
 'upro': ['юнипро', 'upro', 'unip', 'эон россия', 'березовская грэс'],
 'ucss': ['окс',
  'ucss',
  'united credit',
  'объединенные кредитн',
  'ru000a0jtqk0'],
 'trmk': ['тмк', 'трубная', 'trmk', 'tmk', 'трмк'],
 'tcsg': ['tcs', 'tinkof', 'тисиэс', 'тинек', 'тинько', 'us87238u2033'],
 'tatn': ['татнефт', 'tatneft', 'tatn', 'маганов', 'ru0009033591'],
 'tanl': ['tanl', 'tantal', 'тант'],
 'sngs': ['сургут',
  'sngs',
  'surgutneftega',
  'кубышка',
  'сурик',
  'снгс',
  'ru0008926258'],
 'smlt': ['самолет', 'smlt', 'samolet', 'ru000a0zzg02'],
 'sber': ['sber', 'сбер', 'греф', 'ru0009029540'],
 'rual': ['русал', 'rusal', 'rual', 'русский алюминий', 'краз'],
 'rtkm': ['rtkm', 'rostelecom', 'ростел', 'tele2', 'ртк', 'рткм', 'осеевский'],
 'rsti': ['россет', 'rsti', 'российские сети', 'ru000a0jpvj0'],
 'rosn': ['rosn', 'роснефт', 'сечин', 'ru000a0j2q06'],
 'rlmn': ['ролл', 'ролман', 'rlmn', 'роля'],
 'rbcm': ['рбк', 'rbcm', 'rbk', 'rbc', 'ru000a0jr6a6'],
 'rasp': ['rasp', 'распад'],
 'qiwi': ['qiwi', 'киви'],
 'plzl': ['plzl', 'polyus', 'полюс', 'polyusgold', 'сухой лог'],
 'pikk': ['пик', 'pik', 'мортон'],
 'phor': ['фосагр', 'phor', 'phosagr', 'фосфат', 'dap'],
 'paza': ['павлавт',
  'paza',
  'pavlovo',
  'павловский авт',
  'pavlovskiy avt',
  'паз',
  'ru0009083240'],
 'nvtk': ['nvtk', 'новатэк', 'нова тэк', 'novatek', 'ямал спг', 'михельсон'],
 'nmtp': ['нмтп', 'nmtp', 'новороссийск', 'новопорт'],
 'nlmk': ['нлмк', 'nlmk', 'липецк', 'лисин', 'липк'],
 'nknc': ['nknc', 'нижнекамск', 'нкнх', 'нефтехим'],
 'mvid': ['мвидео',
  'mvid',
  'м.видео',
  'м видео',
  'эльдорадо',
  'изосимов',
  'ru000a0jpga0'],
 'mtss': ['mts', 'мтс'],
 'mtlr': ['мечел', 'mtlr', 'mechel', 'ru000a0jpv70'],
 'moex': ['московская биржа', 'мосбирж', 'moex', 'москухн', 'ru000a0jr4a1'],
 'mgnt': ['mgnt', 'магнит', 'magnit', 'марафон групп', 'галицкий', 'тандер'],
 'mail': ['mail',
  'mail.ru',
  'mail',
  'мэйл.ру',
  'мэйл ру',
  'добродеев',
  'us5603172082'],
 'magn': ['ммк', 'magn', 'магнитк', 'магнитогорск', 'магна'],
 'lsrg': ['лср', 'lsr', 'элэсэр'],
 'lsng': ['ленэнерг', 'lsng', 'ленк', 'петербургские электросети'],
 'lnta': ['лент', 'lent', 'lnta', 'tpg capital'],
 'lkoh': ['lkoh',
  'luk',
  'алекперов',
  'лукойл',
  'лучок',
  'лучк',
  'ru0009024277'],
 'krot': ['красокт',
  'krot',
  'крот',
  'krasnyj octyabr',
  'красный октябрь',
  'krasny oktyabr',
  'ru0008913850'],
 'kmaz': ['камаз', 'kmaz', 'kamaz', 'ru0008959580'],
 'iskj': ['искч',
  'институт стволовых клеток',
  'стволовые клетки',
  'стволовых клеток',
  'iskj'],
 'irao': ['irao',
  'inter rao',
  'борис ковальчук',
  'ru000a0jpnm1',
  'рао',
  'интер'],
 'hydr': ['русгидр', 'hydr', 'rushydr', 'гидра', 'ru000a0jpkh7'],
 'hals': ['галс', 'hals', 'пучков', 'ru000a0jnp96'],
 'gmkn': ['gmkn',
  'норникел',
  'nornickel',
  'потанин',
  'гмк',
  'норк',
  'интеррос',
  'гамак',
  'ru0007288411'],
 'gche': ['черкизов', 'cherkizov', 'che'],
 'gazp': ['gazp',
  'газпром',
  'миллер',
  'газик',
  'нашевсе',
  'северный поток',
  'южный поток',
  'сила сибири',
  'ru0007661625'],
 'five': ['five',
  'x5',
  'пятерочк',
  'шехтерман',
  'перекрест',
  'карусел',
  'us98387e2054'],
 'enru': ['энелрос', 'enru', 'enel', 'энел', 'ru000a0f5un3'],
 'dsky': ['dsky', 'detskiy mir', 'детский', 'ru000a0jsq90'],
 'chmf': ['северстал', 'chmf', 'severstal', 'севка', 'череповец', 'мордашов'],
 'chgz': ['рн-запсиб',
  'рн запсиб',
  'chgz',
  'rn-western siberia',
  'рн-западная сибирь',
  'буй',
  'ru0009082291'],
 'belu': ['белуг', 'belu', 'belug', 'ru000a0hl5m1'],
 'bane': ['башнефт', 'bane', 'bashneft', 'ru0007976957'],
 'aptk': ['aptk', 'аптек', '36 и 6', '36.6'],
 'amez': ['amez', 'ашинский', 'аша', 'amet', 'амез', 'амет'],
 'alrs': ['алрос',
  'alrosa',
  'alrs',
  'алмаз',
  'бриллиант',
  'нюрб',
  'ru0007252813'],
 'aflt': ['аэрофлот',
  'aflt',
  'добролет',
  'аврор',
  'российские авиалинии',
  'полубояринов'],
 'afks': ['афк', 'afks', 'afk sistem', 'ozon', 'евтушенков', 'систем']}

### Загружаем сообщения из чатов
Парсились все сообщения в данных чатах. Данные актуальны на 15.03.2021

Файл - ссылка - название - число участников

chat25 - t.me/fondrynok - 'Фондовый рынок Чат №1' - 2741

chat26 - t.me/cbrstocks_chat - 'Сигналы РЦБ [ЧАТ]' - 3529

chat27 - t.me/AK47PFLCHAT - 'РынкиДеньгиВласть чат' - 6445

chat28 - t.me/invest_future_chat - 'InvestFuture Chat' - 2097

chat48 - t.me/Profit_chat24 - 'ProfitGate чат' - 1598

df25 = pd.read_csv('/is/texts/tg/orig/chat25.csv', sep='|')
df26 = pd.read_csv('/is/texts/tg/orig/chat26.csv', sep='|')
df27 = pd.read_csv('/is/texts/tg/orig/chat27.csv', sep='|')
df28 = pd.read_csv('/is/texts/tg/orig/chat28.csv', sep='|')
df48 = pd.read_csv('/is/texts/tg/orig/chat48.csv', sep='|')

df25['src'] = 25
df26['src'] = 26
df27['src'] = 27
df28['src'] = 28
df48['src'] = 48

df = pd.concat([df25, df26, df27, df28, df48])
df.shape

df = df[['date', 'message', 'src']]
df.head()

df.to_csv('/is/texts/tg/text.csv', sep='|', index=False)

df = pd.read_csv('/is/texts/tg/text.csv', sep='|')
df.head()

print(df.shape)
df.drop_duplicates(inplace=True)
print(df.shape)

df['lower_text'] = df.message.str.lower()
df.head()

#### Удаляем перенос строки

df = df.replace('\n',' ', regex=True)
df.head()

#### Меняем ё на е

df['lower_text'] = df['lower_text'].str.replace('ё', 'е')

print(df.shape)
df.drop_duplicates('lower_text', inplace=True)
print(df.shape)

df.to_csv('/is/texts/tg/text_lower.csv', sep='|', index=False)

### Находим сообщения о нужных компаниях

df = pd.read_csv('/is/texts/tg/text_lower.csv', sep='|')
df.shape

852065//51000

def find_ticker(df, tickers, field):
    
    '''
    Функция принимает датафрейм, 
    название тикера, который представляет список с альтернативными наименованиями компании тикера
    и название текстового столбца в котором будет происходить поиск данного тикера.
    '''
    
    df[field] = df[field].astype(str) 
           
    for ticker in tqdm(tickers, position=0, leave=True):
        df[ticker] = 0
        for i in range(df.shape[0]): 
            for alias in tickers[ticker]:
                if alias in df[field].iloc[i]:
                    df[ticker].iloc[i] = 1
                    break

out.shape

# Делаем частями, из-за трудозатрат

#out = pd.DataFrame()
low = 0
up = 51000
for i in range(16):
    print(low, up)    
    df_tic = df.iloc[low:up]
    find_ticker(df_tic, tickers, 'lower_text')
 
    low = up
    up += 51000
    out = pd.concat([out, df_tic])
    out.to_csv(f'/is/texts/tg/text_lower_ticker_{i}.csv', sep='|', index=False)
    
out.to_csv('/is/texts/tg/text_lower_ticker.csv', sep='|', index=False)

out.to_csv('/is/texts/tg/text_lower_ticker.csv', sep='|', index=False)

#### Объединяем упоминания в 1 столбец, чтобы убрать новости с 2 и более разными компаниями

for i in tickers.keys():
    print(f"out['{i}']", end=' + ', sep='')

out['flag'] = out['vtbr'] + out['zvez'] + out['zill'] + out['yndx'] + out['uwgn'] + out['upro'] + out['ucss'] + out['trmk'] + out['tcsg'] + out['tatn'] + out['tanl'] + out['sngs'] + out['smlt'] + out['sber'] + out['rual'] + out['rtkm'] + out['rsti'] + out['rosn'] + out['rlmn'] + out['rbcm'] + out['rasp'] + out['qiwi'] + out['plzl'] + out['pikk'] + out['phor'] + out['paza'] + out['nvtk'] + out['nmtp'] + out['nlmk'] + out['nknc'] + out['mvid'] + out['mtss'] + out['mtlr'] + out['moex'] + out['mgnt'] + out['mail'] + out['magn'] + out['lsrg'] + out['lsng'] + out['lnta'] + out['lkoh'] + out['krot'] + out['kmaz'] + out['iskj'] + out['irao'] + out['hydr'] + out['hals'] + out['gmkn'] + out['gche'] + out['gazp'] + out['five'] + out['enru'] + out['dsky'] + out['chmf'] + out['chgz'] + out['belu'] + out['bane'] + out['aptk'] + out['amez'] + out['alrs'] + out['aflt'] + out['afks']

#### Оставляем те новости у которых есть упоминания компаний и удяляем лишние поля

print(out.shape)
out = out[out.flag == 1]
print(out.shape)

out.head()

out.to_csv('/is/texts/tg/text_lower_ticker_flag.csv', sep='|', index=False)

#### Добавление тикеров вместо 1 и 0

for i in tickers:
    print(f'''out['{i}'] = out['{i}'].astype(str)
out['{i}'] = out['{i}'].str.replace('1.0','{i}').replace('0.0','')
''')

for i in tickers:
    print(f'''out['{i}'] = out['{i}'].astype(str).str.replace('1','{i}').replace('0','')''')

out['vtbr'] = out['vtbr'].astype(str).str.replace('1','vtbr').replace('0','')
out['zvez'] = out['zvez'].astype(str).str.replace('1','zvez').replace('0','')
out['zill'] = out['zill'].astype(str).str.replace('1','zill').replace('0','')
out['yndx'] = out['yndx'].astype(str).str.replace('1','yndx').replace('0','')
out['uwgn'] = out['uwgn'].astype(str).str.replace('1','uwgn').replace('0','')
out['upro'] = out['upro'].astype(str).str.replace('1','upro').replace('0','')
out['ucss'] = out['ucss'].astype(str).str.replace('1','ucss').replace('0','')
out['trmk'] = out['trmk'].astype(str).str.replace('1','trmk').replace('0','')
out['tcsg'] = out['tcsg'].astype(str).str.replace('1','tcsg').replace('0','')
out['tatn'] = out['tatn'].astype(str).str.replace('1','tatn').replace('0','')
out['tanl'] = out['tanl'].astype(str).str.replace('1','tanl').replace('0','')
out['sngs'] = out['sngs'].astype(str).str.replace('1','sngs').replace('0','')
out['smlt'] = out['smlt'].astype(str).str.replace('1','smlt').replace('0','')
out['sber'] = out['sber'].astype(str).str.replace('1','sber').replace('0','')
out['rual'] = out['rual'].astype(str).str.replace('1','rual').replace('0','')
out['rtkm'] = out['rtkm'].astype(str).str.replace('1','rtkm').replace('0','')
out['rsti'] = out['rsti'].astype(str).str.replace('1','rsti').replace('0','')
out['rosn'] = out['rosn'].astype(str).str.replace('1','rosn').replace('0','')
out['rlmn'] = out['rlmn'].astype(str).str.replace('1','rlmn').replace('0','')
out['rbcm'] = out['rbcm'].astype(str).str.replace('1','rbcm').replace('0','')
out['rasp'] = out['rasp'].astype(str).str.replace('1','rasp').replace('0','')
out['qiwi'] = out['qiwi'].astype(str).str.replace('1','qiwi').replace('0','')
out['plzl'] = out['plzl'].astype(str).str.replace('1','plzl').replace('0','')
out['pikk'] = out['pikk'].astype(str).str.replace('1','pikk').replace('0','')
out['phor'] = out['phor'].astype(str).str.replace('1','phor').replace('0','')
out['paza'] = out['paza'].astype(str).str.replace('1','paza').replace('0','')
out['nvtk'] = out['nvtk'].astype(str).str.replace('1','nvtk').replace('0','')
out['nmtp'] = out['nmtp'].astype(str).str.replace('1','nmtp').replace('0','')
out['nlmk'] = out['nlmk'].astype(str).str.replace('1','nlmk').replace('0','')
out['nknc'] = out['nknc'].astype(str).str.replace('1','nknc').replace('0','')
out['mvid'] = out['mvid'].astype(str).str.replace('1','mvid').replace('0','')
out['mtss'] = out['mtss'].astype(str).str.replace('1','mtss').replace('0','')
out['mtlr'] = out['mtlr'].astype(str).str.replace('1','mtlr').replace('0','')
out['moex'] = out['moex'].astype(str).str.replace('1','moex').replace('0','')
out['mgnt'] = out['mgnt'].astype(str).str.replace('1','mgnt').replace('0','')
out['mail'] = out['mail'].astype(str).str.replace('1','mail').replace('0','')
out['magn'] = out['magn'].astype(str).str.replace('1','magn').replace('0','')
out['lsrg'] = out['lsrg'].astype(str).str.replace('1','lsrg').replace('0','')
out['lsng'] = out['lsng'].astype(str).str.replace('1','lsng').replace('0','')
out['lnta'] = out['lnta'].astype(str).str.replace('1','lnta').replace('0','')
out['lkoh'] = out['lkoh'].astype(str).str.replace('1','lkoh').replace('0','')
out['krot'] = out['krot'].astype(str).str.replace('1','krot').replace('0','')
out['kmaz'] = out['kmaz'].astype(str).str.replace('1','kmaz').replace('0','')
out['iskj'] = out['iskj'].astype(str).str.replace('1','iskj').replace('0','')
out['irao'] = out['irao'].astype(str).str.replace('1','irao').replace('0','')
out['hydr'] = out['hydr'].astype(str).str.replace('1','hydr').replace('0','')
out['hals'] = out['hals'].astype(str).str.replace('1','hals').replace('0','')
out['gmkn'] = out['gmkn'].astype(str).str.replace('1','gmkn').replace('0','')
out['gche'] = out['gche'].astype(str).str.replace('1','gche').replace('0','')
out['gazp'] = out['gazp'].astype(str).str.replace('1','gazp').replace('0','')
out['five'] = out['five'].astype(str).str.replace('1','five').replace('0','')
out['enru'] = out['enru'].astype(str).str.replace('1','enru').replace('0','')
out['dsky'] = out['dsky'].astype(str).str.replace('1','dsky').replace('0','')
out['chmf'] = out['chmf'].astype(str).str.replace('1','chmf').replace('0','')
out['chgz'] = out['chgz'].astype(str).str.replace('1','chgz').replace('0','')
out['belu'] = out['belu'].astype(str).str.replace('1','belu').replace('0','')
out['bane'] = out['bane'].astype(str).str.replace('1','bane').replace('0','')
out['aptk'] = out['aptk'].astype(str).str.replace('1','aptk').replace('0','')
out['amez'] = out['amez'].astype(str).str.replace('1','amez').replace('0','')
out['alrs'] = out['alrs'].astype(str).str.replace('1','alrs').replace('0','')
out['aflt'] = out['aflt'].astype(str).str.replace('1','aflt').replace('0','')
out['afks'] = out['afks'].astype(str).str.replace('1','afks').replace('0','')

out.head()

for i in tickers:
    print(f"'{i}', ", sep='', end='')

out['ticker'] = out[['vtbr', 'zvez', 'zill', 'yndx', 'uwgn', 'upro', 'ucss', 'trmk', 'tcsg', 'tatn', 'tanl', \
                 'sngs', 'smlt', 'sber', 'rual', 'rtkm', 'rsti', 'rosn', 'rlmn', 'rbcm', 'rasp', 'qiwi', \
                 'plzl', 'pikk', 'phor', 'paza', 'nvtk', 'nmtp', 'nlmk', 'nknc', 'mvid', 'mtss', 'mtlr', \
                 'moex', 'mgnt', 'mail', 'magn', 'lsrg', 'lsng', 'lnta', 'lkoh', 'krot', 'kmaz', 'iskj', \
                 'irao', 'hydr', 'hals', 'gmkn', 'gche', 'gazp', 'five', 'enru', 'dsky', 'chmf', 'chgz', \
                 'belu', 'bane', 'aptk', 'amez', 'alrs', 'aflt', 'afks']].agg(''.join, axis=1)

out.head()

out = out[['date','message','src','lower_text','ticker']]
out.head()

### Работа с датой
0 - Пн

datetimes = pd.to_datetime(out['date'])
out['day'] = datetimes.dt.day
out['month'] = datetimes.dt.month
out['year'] = datetimes.dt.year
out['day_of_week'] = datetimes.dt.dayofweek
out.head()

out.year.unique()

out['dt'] = (out.year.astype(str) + out.date.astype(str).str[5:7] + out.date.astype(str).str[8:10]).astype(int)
out.head()

out.dt, out.date = out.date, out.dt
out.head()

out.date = out.date.astype(float)

out.to_csv('/is/texts/tg/text_lower_ticker_flag_dt.csv', sep='|', index=False)

#### Сохраняем даты для джоина новостей и цен

out[['ticker', 'date']].drop_duplicates().to_csv('/is/texts/tg/news_dates.csv', sep='|', index=False)
